requirements.txt
asgiref==3.8.1
Django==5.1.4
django-recaptcha==4.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg==3.2.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
